OpenTSDB: Run suggest queries through the data source backend
